CodeStyle III

+++ core/src/de/dhbw/satp/main/FinalStatics.java
@@ -2,13 +2,9 @@ package de.dhbw.satp.main;
 
 public class FinalStatics {
 
-    //Object
     public static final float BLOCK_UNIT = 16f;
-
     public static final float VIRTUAL_WIDTH = 25 * BLOCK_UNIT;
     public static final float VIRTUAL_HEIGHT = 13 * BLOCK_UNIT;
-
-    public static final int PPM = 100;
-
+    public static final float PPM = 100;
 
 }
